requirements.txt
streamlit
scikit-learn
python-dotenv
google-genai
matplotlib
reportlab
bcrypt